refactor(probes): Phase 3.1 — extract probe_* to backend/probes/
Pure move. _http_json, _openai_models, probe_ollama/tgi/koboldcpp/invokeai/
a1111/whisper_asr/triton/wyoming/comfy, PROBES, _match_probe, _match_probe_key,
CATALOG_MAX, probe_models, probe_host_metrics, probe_host, probe_http,
probe_tcp all moved to backend/probes/__init__.py. app.py re-exports all
moved names. Snapshot suite confirms API responses unchanged.
